Fixed API 9 support and some layout changes
Running the app on API 10 crashed due to a layout file using an
attribute that exists only in API 11 and above. This was fixed by
creating a copy in values-v11 and removing the attribute in the
original file.

MainActivity now has drop down menu to change the default sort for
submissions.

Successive user login/logout doesn’t create a clean blank state and
seems to maintain some session information. This hasn’t been completely
fixed yet (sometimes user needs to completely close the app). This is
likely linked to the Apache HttpClient, but further investigation is
necessary.